backend/requirements.txt
asgiref==3.7.2
boto3==1.34.93
botocore==1.34.93
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-storages==1.14.2
djangorestframework==3.15.0
jmespath==1.0.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==1.26.18
